Move CyclomaticComplexity from metrics to design w/ deprecated property
