--- pmd-java/src/main/resources/rulesets/java/metrics.xml
@@ -1,66 +1,0 @@
-<?xml version="1.0"?>
-
-<ruleset name="Metrics temporary ruleset"
-         xmlns="http://pmd.sourceforge.net/ruleset/2.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://pmd.sourceforge.net/ruleset/2.0.0 http://pmd.sourceforge.net/ruleset_2_0_0.xsd">
-
-    <description>
-        These are rules which use the Metrics Framework to calculate metrics.
-    </description>
-
-    <rule name="CyclomaticComplexity"
-          message="The {0} ''{1}'' has a{2} cyclomatic complexity of {3}."
-          since="1.03"
-          class="net.sourceforge.pmd.lang.java.metrics.rule.CyclomaticComplexityRule"
-          externalInfoUrl="${pmd.website.baseurl}/pmd_rules_java_metrics.html#cyclomaticcomplexity">
-        <description><![CDATA[
-            The complexity of methods directly affects maintenance costs and readability. Concentrating too much decisional logic
-            in a single method makes its behaviour hard to read and change.
-
-            Cyclomatic complexity assesses the complexity of a method by counting the number of decision points in a method,
-            plus one for the method entry. Decision points are places where the control flow jumps to another place in the
-            program. As such, they include all control flow statements, such as `if`, `while`, `for`, and `case`. For more
-            details on the calculation, see the documentation of the [Cyclo metric](/pmd_java_metrics_index.html#cyclomatic-complexity-cyclo).
-
-            Generally, numbers ranging from 1-4 denote low complexity, 5-7 denote moderate complexity, 8-10 denote
-            high complexity, and 11+ is very high complexity. By default, this rule reports methods with a complexity >= 10.
-            Additionnally, classes with many methods of moderate complexity get reported as well once the total of their
-            methods' complexities reaches 80, even if none of the methods was directly reported.
-
-            Reported methods should be broken down into several smaller methods. Reported classes should probably be broken down
-            into subcomponents.]]>
-        </description>
-        <priority>3</priority>
-        <example>
-            <![CDATA[
-class Foo {
-  void baseCyclo() {                // Cyclo = 1
-    highCyclo();
-  }
-
-  void highCyclo() {                // Cyclo = 10: reported!
-    int x = 0, y = 2;
-    boolean a = false, b = true;
-
-    if (a && (y == 1 ? b : true)) { // +3
-      if (y == x) {                 // +1
-        while (true) {              // +1
-          if (x++ < 20) {           // +1
-            break;                  // +1
-          }
-        }
-      } else if (y == t && !d) {    // +2
-        x = a ? y : x;              // +1
-      } else {
-        x = 2;
-      }
-    }
-  }
-}
-]]>
-        </example>
-    </rule>
-
-
-</ruleset>
